Reverts vllm-project/vllm#3685

This impacted the performance of vLLM.

this commit cause the L4 opt benchmark to drop performance
before: https://buildkite.com/vllm/ci/builds/3695
after: https://buildkite.com/vllm/ci/builds/3716

before
Avg latency: 0.446163778666687 seconds
Throughput: 21.25 requests/s, 10879.19 tokens/s

after
Avg latency: 4.369133323666698 seconds
Throughput: 0.93 requests/s, 474.10 tokens/s